Realizado abstração do controller.
